% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_div.R
\name{age_div}
\alias{age_div}
\title{Divides an individual into parts}
\usage{
age_div(data, show_res = F, extra_var)
}
\arguments{
\item{data}{A data frame with individuals' age-at-death estimations.
Data frame must contain at least two columns:
\code{Age_min}, lower limit of age-at-death interval;
\code{Age_max}, upper limit of age-at-death interval.
For example, the value of 20 means that an individual died
before her/his 20th birthday, i.e. up to 19.9 years of age.
Two other columns are expected (\code{Site} and \code{Culture}), but they are not required.}

\item{show_res}{If \code{TRUE}, results are printed. Defaults to \code{FALSE}.}

\item{extra_var}{Extra columns/variables stored in \code{data},
which should be added to the results.}
}
\value{
A data frame containing individuals splitted into age-at-death categories.

Variables are as follows:
\itemize{
\item \code{Site, Culture, Age-min, Age_max}: Summary of inputs.
If \code{Site} and \code{Culture} are not given in the \code{data},
then \code{Unknown} values are used.
\item \code{A_0, ..., A_99}: Proportion of the individual
belonging to the age-at-death categories betweeen 0 and 99 years. Row sum equals to 1 for each individual.
}
}
\description{
Proportionally splits an individual into 100 age-at-death categories (0-99 years)
on the basis of the estimation of its age-at-death.
}
\examples{
age_div(BAraw, show_res = T) \%>\%
as_tibble()
}
\keyword{age-at-death}
