% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BA}
\alias{BA}
\title{Bocquet-Appel's (2002, Table 1) dataset in the summary data format}
\format{
A tibble with 68 rows and 12 variables:
\describe{
\item{Site}{The name of the cemetery sample}
\item{Front}{Absolute C14 date of the Neolithic transition in the region of the Site}
\item{C14}{Absolute C14 date of the Site}
\item{dt}{The number of years from the Neolithic front, C14-Front}
\item{D0_4}{D0-4; number of skeletons between 0 and 4 years of age}
\item{D5_19}{D5-19; number of skeletons between 5 and 19 years of age}
\item{D20_}{D20+; number of skeletons over 20 years of age}
\item{Total}{Total number of skeletons}
\item{D5_19_D5_}{D5-19/D5+ ratio or index P}
\item{D5_}{D5+; number of skeletons over 5 years of age}
\item{D5_D20_}{D5+/D20+ ratio}
\item{Culture}{Period to which Site belongs}
}
}
\usage{
BA
}
\description{
Bocquet-Appel J-P (2002). Paleoanthropological traces of a Neolithic demographic transition. Current Anthropology 43: 637-650.
Summary of distribution of deaths in 68 Mesolithic and Neolithic European cemetery samples.
Three new variables were added to the dataset (D5_, D5_D20_, and Culture).
The dataset is in the summary data format, each row represents one skeletal sample.
}
\keyword{datasets}
