% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{package}
\name{demrat-package}
\alias{demrat}
\alias{demrat-package}
\title{demrat: Estimation Of Fertility And Growth Rates For Skeletal Samples}
\description{
The package enables the prediction of growth rate, crude birth rate, and total fertility rate for skeletal samples by using age-at-death ratios computed from the number of skeletons in two age-at-death categories. The prediction is based on a regression model derived from a unique set of reference skeletal samples of the same size as the cemetery sample under study.
}
\author{
\strong{Maintainer}: Patrik Galeta \email{galetap@gmail.com} (\href{https://orcid.org/0000-0002-4937-1369}{ORCID})

}
\keyword{internal}
