% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diest.R
\name{diest}
\alias{diest}
\title{Estimates growth and fertility rate for skeletal samples}
\usage{
diest(
  dr_data,
  summary = T,
  pred_level = 0.95,
  sss = T,
  samples = 100,
  e0_min = 18,
  e0_max = 25,
  growth_min = -3,
  growth_max = 3,
  IV = c("All", "D5_D20_", "D3_D20_", "D1_D20_", "P"),
  DV = c("All", "Growth", "TFR", "CBR"),
  extra_var,
  ...
)
}
\arguments{
\item{dr_data}{A data frame with one row containing data for a single cemetery sample.
Data frame must have four columns at minimum:
\code{Site}, \code{Culture}, \code{D20_} (i.e., D20+, number of adult skeletons in a sample)
and at least one out of the following three columns: \code{D1_D20_, D3_D20_, D5_D20_}
(i.e., D1+/D20+, D3+/D20+, D5+/D20+ age-at-death ratios, respectively).}

\item{summary}{If \code{TRUE}, functions returns only estimations.
If \code{FALSE}, function returns full results including model characteristics, reference data, etc.}

\item{pred_level}{Confidence of the prediction, defaults to 0.95.}

\item{sss}{If \code{TRUE}, simulated skeletal samples are created as the reference.
If \code{FALSE}, reference consists from population data,
the algorithm step, in which simulated skeletal samples are created, is skipped.}

\item{samples}{Number of simulated reference skeletal samples to be created.}

\item{e0_min}{Minimum value of life expectancy at birth in population from which skeletal samples are drawn.}

\item{e0_max}{Maximum value of life expectancy at birth in population from which skeletal samples are drawn.}

\item{growth_min}{Minimum value of annual growth rate in population from which skeletal samples are drawn.}

\item{growth_max}{Maximum value of annual growth rate in population from which skeletal samples are drawn.}

\item{IV}{Independent variable (the age-at-death ratio) used in the prediction. One or more items from \code{c("All", "D5_D20_", "D3_D20_", "D1_D20_", "P")}.}

\item{DV}{Dependent (demographic) variable to be estimated. One or more items from \code{c("All", "Growth", "TFR", "CBR")}.}

\item{extra_var}{Extra variables to be joined to results (e.g. description, absolute chronology of the site).
Extra variables must be in \code{dr_data} data frame.}

\item{...}{Other parameters (see \code{simdr_CD} and \code{diest_fn})}
}
\value{
Data frame containing estimation of demographic indicators of population
from which a cemetery sample was drawn.

Summary of the inputs (cemetery sample input data):
\itemize{
\item \strong{\code{Site}}: Name of the cemetery sample.
\item \strong{\code{Culture}}: Archaeological culture to which \code{Site} belongs.
\item \strong{\code{dr_data}}: A copy of input data with the age-at-death ratios
for the \code{Site}.
}

Characteristics of model used in prediction.
\itemize{
\item \strong{\code{DV}}: Dependent variable (estimated demographic indicator).
\item \strong{\code{IV}}: Independent variable (age-at-death ratio).
\item \strong{\code{Ref}}: Data frame containing reference sample from which regression model is computed.
\item \strong{\code{Glance}}: A tibble with basic characteristics of the regression model
used in the prediction.
}

Results of demographic estimation.
\itemize{
\item \strong{\code{Est}}: Estimation of demographic indicator (see \code{DV})
based on age-at-death ratio (see \code{IV}).
\item \strong{\code{Lwr}}: Lower limit of the prediction interval.
\item \strong{\code{Upr}}: Upper limit of the prediction interval.
\item \strong{\code{Ratio_lim}}: 97.5 percentile of the age-at-death ratio
computed in the set of simulated reference skeletal samples.
\item \strong{\code{Ratio_eval}}: \code{Out of limits} if the age-at-death ratio observed in the real skeletal sample is higher than the Ratio_lim.
In this case, the ratio may be biased and estimation may be unreliable.
}
}
\description{
Estimates growth, crude birth and total fertility rates based on age-at-death ratios
(e.g., D5+/D20+) summarising distribution of deaths in a cemetery sample.
Function can be applied to several cemetery samples at once.
}
\examples{
# Demographic estimation based on the original Bocquet-Appel (2002, Table 1) data
BA \%>\%
dplyr::slice(5, 12) \%>\%
diest()

# Demographic estimation based on the reconstructed Bocquet-Appel (2002) raw data
BAraw \%>\%
dr() \%>\%
dplyr::slice(1:2) \%>\%
diest()

# Select predicted demographic variables and predictors that are showed in the results
BA \%>\%
dplyr::slice(12) \%>\%
diest(summary = TRUE, IV=c("D5_D20_"), DV=c("Growth", "TFR"))

# Adding extra variables
BA \%>\%
dplyr::slice(5, 12) \%>\%
diest(extra_var = c(Front, C14, dt),
growth_min = -2, growth_max = 2)
}
\keyword{demographic}
\keyword{estimation}
