% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diest_fn.R
\name{diest_fn}
\alias{diest_fn}
\title{Estimates growth and fertility rate for a single cemetery sample (helper function)}
\usage{
diest_fn(dr_data, glanced = F, pred_level = 0.95, extra_var, ...)
}
\arguments{
\item{dr_data}{A data frame, each row contains data for a single cemetery sample.
Data frame must have two columns at minimum:
\code{D20_} (i.e., D20+, number of adult skeletons in a sample)
and at least one of the following three columns: \code{D1_D20_, D3_D20_, D5_D20_}
(i.e., D1+/D20+, D3+/D20+, D5+/D20+ age-at-death ratios, respectively).}

\item{glanced}{If \code{TRUE}, resulting tibble is glanced.}

\item{pred_level}{Confidence of the prediction, default is 0.95.}

\item{extra_var}{Extra variables to be joined to results (e.g. description, absolute chronology of the site).
Extra variables must be in \code{dr_data} data frame.}

\item{...}{Other parameters (see \code{simdr_CD})}
}
\value{
Data frame containing estimation of demographic indicators of population
from which a cemetery sample was drawn. See \code{diest()} for details.
}
\description{
Estimates growth, crude birth and total fertility rates
based on age-at-death ratios (e.g., D5+/D20+)
computed in cemetery sample.
Helper function, does not show results until they are saved to an R object
}
\examples{
res <-
BAraw \%>\%
dr() \%>\%
slice(1) \%>\%
diest_fn()

res

}
\keyword{demographic}
\keyword{estimation}
