% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.R
\name{dr}
\alias{dr}
\title{Calculates age-at-death ratios}
\usage{
dr(data, extra_var)
}
\arguments{
\item{data}{A data frame with individuals' age-at-death estimations.
Data frame must contain at least two columns:
\code{Age_min}, Lower limit of age-at-death interval;
\code{Age_max}, Upper limit of age-at-death interval.
The interval is right-open. The value of 20 years means that an individual died
before her/his 20th birthday, i.e. up to 19.9 years of age.
Two other columns are expected (\code{Site} and \code{Culture}), but they are not required.}

\item{extra_var}{Extra variables to be joined to results (e.g. description, absolute chronology of the site).
Extra variables must be in \code{data}.}
}
\value{
Data frame summarising dstribution of deaths by skeletal samples.

Summary of the inputs:
\itemize{
\item \strong{\code{Site}}: Name of the cemetery sample.
\item \strong{\code{Culture}}: Archaeological culture to which Site belongs.
}

Number of skeletons on various age-at-death categories:
\itemize{
\item \strong{\code{n}}: Total number of skeletons.
\item \strong{\code{D0}}: Number of newborn deaths.
\item \strong{\code{D1_, D3_, D5_, D15_, and D20_}}: D1+, D3+, D5+, D15+, and D20+;
number of deaths older than 1, 3, 5, 15, and 20 years respectively.
\item \strong{\code{D0_14, D5_14, and D5_19}}: D0-14, D5-14, and D5-19;
number of deaths between 0-14, 5-14 and 5-19 years, respectively.
\item \strong{\code{D1_D20_, D3_D20_, and D5_D20}}: D1+/D20+, D3+/D20+, and D5+/D20+ age-at-death ratios.
\item \strong{\code{JI, and P}}: Bocquet-Appel's juvenility index (D5-14/D20+) and P index (D5-19/D5+).
\item \strong{\code{D0_14_D0_}}: McFadden and Oxenham D0-14/D0+ age-at-death ratio.
}
}
\description{
Based on age-at-death estimation of individuals, the function calculates
number of skeletons in several age-at-death groups (e.g., D1+, D20+)
and six demographic ratios (D1+/D20+, D3+/D20+, D5+/D20+, D5-14/D20+ aka juvenility index,
D5-19/D5+ aka P index, and D0-14/D0+) for skeletal sample(s).
}
\examples{
dr(BAraw)

BAraw \%>\%
select(-Site) \%>\%
dr()
}
\keyword{age-at-death}
\keyword{ratio}
