% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diest.R
\name{plot_diest}
\alias{plot_diest}
\title{Plots prediction model}
\usage{
plot_diest(
  dr_data,
  IV = c("D5_D20_", "D3_D20_", "D1_D20_"),
  DV = c("Growth", "TFR", "CBR"),
  base_size = 9,
  ...
)
}
\arguments{
\item{dr_data}{A data frame with one row containing data for a single cemetery sample.
Data frame must have four columns at minimum:
\code{Site}, \code{Culture}, \code{D20_} (i.e., D20+, number of adult skeletons in a sample)
and at least one out of the following three columns: \code{D1_D20_, D3_D20_, D5_D20_}
(i.e., D1+/D20+, D3+/D20+, D5+/D20+ age-at-death ratios, respectively).
If data frame contains total number of skeletons, its name should be n.}

\item{IV}{Independent variable (the age-at-death ratio) used in the prediction. One or more items from \code{c("All", "D5_D20_", "D3_D20_", "D1_D20_", "P")}.}

\item{DV}{Dependent (demographic) variable to be estimated. One or more items from \code{c("All", "Growth", "TFR", "CBR")}.}

\item{...}{Other parameters (see \code{diest})}
}
\value{
Plot of the unique regression model used in the prediction
of the selected demographic variable from the selected age-at-death ratio.
}
\description{
The function visualizes an unique prediction model
used for a estimation of demographic parameters at a given Site.
In the plot, the age-at-death ratio (independent variable) is on the X axis,
the demographic variable (dependent variable) is on the Y axis.
The scatterplot contains the reference set of simulated skeletal samples, regression line
and the data for a given Site.
Function can be applied to a single skeletal sample only.
}
\examples{
# Default plot (the relationship of Growth on the D5+/D20+ ratio)
BA \%>\%
slice(12) \%>\%
plot_diest()

# Plot of the relationship of TFR on the D5+/D20+ ratio
BA \%>\%
slice(12) \%>\%
plot_diest(DV="TFR")

# Changing parameters of the prediction model
BA \%>\%
slice(12) \%>\%
plot_diest(samples=500, e0_min=25, e0_max=30)
}
\keyword{demographic}
\keyword{estimation}
