% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{W_coef_Est}
\alias{W_coef_Est}
\title{Siler coefficients for Coale and Demeny (1983) life tables}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 63 rows and 7 columns.
}
\usage{
W_coef_Est
}
\description{
Siler coefficients for Coale and Demeny (1983) life tables
}
\keyword{datasets}
