% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdr_CD.R
\name{simdr_CD}
\alias{simdr_CD}
\title{Create a reference set of simulated skeletal samples}
\usage{
simdr_CD(
  sss = T,
  samples = 100,
  D20_raw = 50,
  e0_min = 20,
  e0_max = 30,
  growth_min = -3,
  growth_max = 3,
  ...
)
}
\arguments{
\item{sss}{If \code{TRUE}, simulated skeletal samples are created as the reference.
If \code{FALSE}, reference consists from population data,
the step 3, in which simulated skeletal samples are created, is skipped.}

\item{samples}{Number of simulated skeletal samples to be created.}

\item{D20_raw}{D20+, number of adult skeletons (20+ years) in simulated skeletal samples.}

\item{e0_min}{Minimum value of life expectancy at birth in population from which skeletal samples are drawn.}

\item{e0_max}{Maximum value of life expectancy at birth in population from which skeletal samples are drawn.}

\item{growth_min}{Minimum value of annual growth rate in population from which skeletal samples are drawn.}

\item{growth_max}{Maximum value of annual growth rate in population from which skeletal samples are drawn.}

\item{extra_var}{Extra variables to be joined to results (e.g. description, absolute chronology of the site).
Extra variables must be in \code{dr_data} data frame.}
}
\value{
A data frame with summary results.

Summary of function inputs:
\itemize{sss
\item \strong{\code{SSS}}: Were simulated skeletal samples generated (\code{TRUE} or \code{FALSE})?
See \code{sss} argument.
\item \strong{\code{Samples}}: Total number of skeletal samples created (number of rows in results).
\item \strong{\code{D20_raw}}: D20+, number of adult skeletons (20+ years) in simulated skeletal samples.
If \code{sss} is \code{FALSE}, \code{D20_raw} is \code{NA}.
\item \strong{\code{e0_min}}: Minimum value of life expectancy at birth of population from which skeletal samples were drawn.
\item \strong{\code{e0_max}}: Maximum value of life expectancy at birth of population from which skeletal samples were drawn.
\item \strong{\code{growth_min}}: Minimum value of annual growth rate of population from which skeletal samples were drawn.
\item \strong{\code{growth_max}}: Maximum value of annual growth rate of population from which skeletal samples were drawn.
\item \strong{\code{Alpha, Beta}}: Not applicable.
}

Demographic characteristic of reference population.
If \code{sss} is TRUE, this is a population from which a skeletal sample is drawn:
\itemize{
\item \strong{\code{Growth}}: Annual growth rate randomly drawn
from interval between \code{growth_min} and \code{growth_max}.
\item \strong{\code{l27.5}}: The number of persons surviving to exact age 27.5,
calculated from a life table.
\item \strong{\code{e0}}: Life expectancy at birth.
\item \strong{\code{CBR}}: Crude birth rate.
\item \strong{\code{TFR}}:\code{ }Total fertility rate.
}

Characteristics of distribution of deaths
(number of deaths in the respective age-at-death interval and demographic ratios).
\itemize{
\item \strong{\code{n}}: Total number of skeleton in skeletal sample
\item \strong{\code{D0, D1_, ..., D5_19}: Number of individuals in age-at-death groups;
D0+, D1+, ..., D5-19.
\item \strong{\code{D1_D20_, ..., D5_D20_}}: Age-at-death ratios; D1+/D20+, ..., D5+/D20+.
\item \strong{\code{JI}}: Juvenility index (D5-14/D20+).
\item \strong{\code{P index}}: D5-19/D5+.
\item \strong{\code{D0_14_D0_}}: D0-14/D0+ ratio.
}
}
\description{
Step 1: Function calculate life table of population based on Coale and Demeny West model.
Step 2: It applies a growth rate and calculate distribution of deaths in stable population.
Step 3: It randomly draws deaths from that distribution of deaths.
Step 4: It calculates age-at-death ratios from the distribution of deaths
(step 2 or 3, it depends on \code{sss} argument.
All three steps are iterated many times to obtain the reference dataset of simulated skeletal samples.
}
\examples{
simdr_CD()

simdr_CD(sss=F)
}
\keyword{simulation}
